+++ windows/AudioRouter.Windows/pch.h
@@ -1,17 +1,8 @@
 #pragma once
 
 #include <unknwn.h>
+#include <winrt/base.h>
 #include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Foundation.Collections.h>
-#include <winrt/Microsoft.UI.Xaml.h>
-#include <winrt/Microsoft.UI.Xaml.Controls.h>
-#include <winrt/Microsoft.UI.Xaml.Controls.Primitives.h>
-#include <winrt/Microsoft.UI.Xaml.Data.h>
-#include <winrt/Microsoft.UI.Xaml.Input.h>
-#include <winrt/Microsoft.UI.Xaml.Markup.h>
-#include <winrt/Microsoft.UI.Xaml.Media.h>
-#include <winrt/Microsoft.UI.Xaml.Navigation.h>
-#include <winrt/Microsoft.UI.Dispatching.h>
 #include <winrt/Windows.ApplicationModel.Activation.h>
 #include <winrt/Windows.System.h>
 #include <winrt/Windows.UI.h>
@@ -30,6 +21,7 @@
 #include <algorithm>
 #include <chrono>
 #include <cmath>
+#include <cstdlib>
 #include <cwctype>
 #include <filesystem>
 #include <functional>
@@ -41,7 +33,3 @@
 #include <string_view>
 #include <utility>
 #include <vector>
-
-namespace muxc = winrt::Microsoft::UI::Xaml::Controls;
-namespace muxcp = winrt::Microsoft::UI::Xaml::Controls::Primitives;
-namespace mux = winrt::Microsoft::UI::Xaml;